requirements.txt
altair==5.4.1
attrs==24.2.0
azure-core==1.30.2
azure-storage-blob==12.22.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
fonttools==4.53.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
Levenshtein==0.26.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mutagen==1.47.0
narwhals==1.6.2
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.1
pywin32==306
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
